[FEATURE] Prevent connection on driver initialization
Driver is initialized with every call to TYPO3. Even if no FTP resource
is included in the page, the connection gets initialized. This patch
changes the behavior, so the connection is only initialized if an action
needs the connection stream. Furthermore the connection is tested for
each configuration once to validate driver configuration.